Split BugSplatErrorHandler into BugSplatErrorHandler, BugSplat and BugSplatConfig files. Updated index.ts, bugsplat.modules.ts and karma.conf.js with these path changes.
