+++ index.ts
@@ -1,1 +1,3 @@
-export { BugSplatErrorHandler, BugSplat, BugSplatConfig, BugSplatConfigToken } from "./src/bugsplat-error-handler";
+export { BugSplatErrorHandler } from "./src/bugsplat-error-handler";
+export { BugSplat } from "./src/bugsplat";
+export { BugSplatConfig, BugSplatConfigToken } from "./src/bugsplat-config";
