+++ src/bugsplat.module.ts
@@ -2,7 +2,9 @@
 import { BrowserModule } from '@angular/platform-browser';
 import { NgModule, ErrorHandler } from '@angular/core';
 import { HttpClientModule } from "@angular/common/http";
-import { BugSplatErrorHandler, BugSplat, BugSplatConfig } from "./bugsplat-error-handler";
+import { BugSplatErrorHandler } from "./bugsplat-error-handler";
+import { BugSplat } from "./bugsplat";
+import { BugSplatConfig } from "./bugsplat-config";
 
 @NgModule({
     imports: [
